% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrenergycore-package.R
\docType{package}
\name{mrenergycore-package}
\alias{mrenergycore}
\alias{mrenergycore-package}
\title{mrenergycore: Fundamental REMIND preprocessing functions}
\description{
Provides useful functions and a common structure to all the input data required to run the REMIND of model input data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrenergycore}
  \item Report bugs at \url{https://github.com/pik-piam/mrenergycore/issues}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Falk Benke \email{benke@pik-potsdam.de}
}

}
